refactor: split download store schema migrations
